Basic version of arrow smart selection
Up and down arrows still don't work they way I'd expect them to inside
nested inline-exprs. Instead of just going to the closest parent when
the up-arrow is pressed in a nested inline-expr they should go the
left sibling of the first non-inline parent (basically the line above it)
Similarly for the down-arrow.

The materialized underline layouts inside broken-up calls should not be
inline either.