Move drop handling into ExprView in preparation for adding parameters
This has many advantages, we do not need to expose the full ExprArea
to the Editor and both halves of the drag-and-drop code now live in
ExprView. Additionally we can use the useDrop hook instead of manually
manging our listener.

We get rid of 'frozen' expressions, instead introducing a host of new
boolean ExprView properties to describe its behaviour during
drag-and-drop.

* Add a lot more JSDoc comments to ExprView
* Destructure props earlier in the ExprView
* Simplify drop acceptance to just return a boolean